payment-request/PaymentMethodChangeEvent/methodDetails-attribute.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test for PaymentMethodChangeEvent.methodDetails attribute</title>
<link rel="help" href="https://w3c.github.io/browser-payment-api/#dom-paymentmethodchangeevent-methoddetails">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
"use strict";
test(() => {
  const event = new PaymentMethodChangeEvent("test", {
    methodName: "wpt-test",
    methodDetails: {
      test: "pass"
    }
  });
  assert_idl_attribute(event, "methodDetails");
  const { test } = event.methodDetails;
  assert_equals(test, "pass");
}, "Must have a methodDetails IDL attribute, which is initialized with to the methodName dictionary value");
